Changed datatype of memory tracking to size_t

+++ src/src_sharpSAT/Basics.h
@@ -25,7 +25,7 @@ public:
 
     static unsigned int secsTimeBound;
 
-    static unsigned int maxCacheSize;   // maximum Cache Size in bytes
+    static size_t maxCacheSize;   // maximum Cache Size in bytes
     
     static int nodeCount; // Nodes currently in use
